Move specs to docs package
